WIP: crypto: Refactor to move files out of the top level directory
Currently the top level directory contains basically all of the code
for the crypto package. This should be moved into sub-directories.
There are a couple of open questions, how do we want to do a symmetric
interface, and how do we want to structure the encoding. This PR
currently names the encoding package "amino", but this has a clear
naming conflict and will probably be confusing.

This commit doesn't update all of tendermint yet, hence the failing
tests.

Closes #1956